+++ src/reducer/action-creators.js
@@ -1,0 +1,10 @@
+import {ActionType} from "./action-types";
+
+const ActionCreator = {
+  getActiveGenre: (genre) => ({
+    type: ActionType.ACTIVE_GENRE,
+    payload: genre,
+  }),
+};
+
+export {ActionCreator};
