+++ src/reducer/action-types.js
@@ -1,0 +1,5 @@
+const ActionType = {
+  ACTIVE_GENRE: `ACTIVE_GENRE`,
+};
+
+export {ActionType};
